Implement MT call handling with a big change
For MT-Call handling we will take a look at the invocation of the SRI and not the stored one from the LU. Respond with a static reply.

The tests makes sure that we first provide the subscriber info and VMSC address. Actually the test just checks that a camelRoutingInfo is included. We also added a test for the second invocation that will respond with the MSRN.

The decision of which operation to use is taken inside the Dialogue class because it is the easiest place to make the decision instead of putting the handling into just one subclass.