Reject v3 sendAuthenticationInfo from the SGSNs
Apparently equipments expect that on a V3 request we answer if we have a USIM in our HLR. Currently we only support GSM sim cards and would respond with a triplet. The SGSNs will then ask us again for the information and again. Simply abort the dialogue and hope it comes back with a V2 request. Add a testcase.