Modify the SQN handling code
Avoid making big jumps on each SAI. Increment it by one only. In case of a re-sync make a bigger jump (make sure the seq updates by one). This means the SQN will always move forward and fast forward on re-sync. This way we don't re-use the same SQN with a different RAND.